+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Memoization for substring generator is done across instances.
 - Memoization for truncator is done across instances.
 - Coder does no longer have memoization.
+- Coder can no longer be instantiated.
 
 ### Fixed
 - Capitalizer discarded empty strings.